test/adjoint: move rotate functions
